requirements.txt
Flask
Flask-Limiter
gunicorn
pytest
pytest-cov
xvfbwrapper
selenium
flask-jwt-extended
Pillow